added germanized setup

+++ provision/vvv-init.sh
@@ -70,31 +70,60 @@ configure_woo() {
     noroot wp plugin activate woocommerce-disable-wc-admin
   fi
 
-  noroot wp language plugin install woocommerce de_DE
-  noroot wp option update woocommerce_tax_total_display itemized
-  noroot wp option update woocommerce_default_country DE
-  noroot wp option update woocommerce_currency EUR
-  noroot wp option update woocommerce_currency_pos right_space
-  noroot wp option update woocommerce_price_thousand_sep .
-  noroot wp option update woocommerce_price_decimal_sep ,
-  noroot wp option update woocommerce_price_num_decimals 2
-  noroot wp option update woocommerce_weight_unit kg
-  noroot wp option update woocommerce_dimension_unit cm
-  noroot wp option update woocommerce_calc_taxes yes
-  noroot wp option update woocommerce_prices_include_tax yes
-  noroot wp option update woocommerce_tax_display_cart incl
-  noroot wp option update woocommerce_tax_display_shop incl
-  noroot wp option update woocommerce_tax_based_on billing
-  noroot wp option update woocommerce_default_customer_address base
+  FORCE_RESET_WP=$(get_config_value 'force_reset_wp' "")
+  
+  if [ ! -z "${FORCE_RESET_WP}" ]; then
+    noroot wp language plugin install woocommerce de_DE
+    noroot wp option update woocommerce_tax_total_display itemized
+    noroot wp option update woocommerce_default_country DE
+    noroot wp option update woocommerce_currency EUR
+    noroot wp option update woocommerce_currency_pos right_space
+    noroot wp option update woocommerce_price_thousand_sep .
+    noroot wp option update woocommerce_price_decimal_sep ,
+    noroot wp option update woocommerce_price_num_decimals 2
+    noroot wp option update woocommerce_weight_unit kg
+    noroot wp option update woocommerce_dimension_unit cm
+    noroot wp option update woocommerce_calc_taxes yes
+    noroot wp option update woocommerce_prices_include_tax yes
+    noroot wp option update woocommerce_tax_display_cart incl
+    noroot wp option update woocommerce_tax_display_shop incl
+    noroot wp option update woocommerce_tax_based_on billing
+    noroot wp option update woocommerce_default_customer_address base
 
-  noroot wp option update woocommerce_store_address "Schillerstraße 36a"
-  noroot wp option update woocommerce_store_city Berlin
-  noroot wp option update woocommerce_store_postcode 12207
-  noroot wp option delete woocommerce_admin_notices
-  noroot wp wc tool run install_pages --user=1
+    noroot wp option update woocommerce_store_address "Schillerstraße 36a"
+    noroot wp option update woocommerce_store_city Berlin
+    noroot wp option update woocommerce_store_postcode 12207
+    noroot wp option delete woocommerce_admin_notices
+    noroot wp wc tool run install_pages --user=1
 
-  echo " * Installing Woo demo content"
-  noroot wp import wp-content/plugins/woocommerce/sample-data/sample_products.xml --authors=skip --user=1
+    echo " * Installing Woo demo content"
+    noroot wp import wp-content/plugins/woocommerce/sample-data/sample_products.xml --authors=skip --user=1
+  fi
+
+  if $(noroot wp plugin is-installed woocommerce-germanized ); then
+    echo " * Configuring Germanized"
+    noroot wp plugin activate woocommerce-germanized
+
+    if [ ! -z "${FORCE_RESET_WP}" ]; then
+      noroot wp import wp-content/plugins/woocommerce-germanized/dummy-data/dummy-data.xml --authors=skip --user=1
+    fi
+
+    if $(noroot wp plugin is-installed woocommerce-germanized-pro ); then
+      noroot wp plugin activate woocommerce-germanized-pro
+    fi
+
+    if $(noroot wp plugin is-installed woocommerce-germanized-shipments ); then
+      noroot wp plugin activate woocommerce-germanized-shipments
+    fi
+
+    if $(noroot wp plugin is-installed woocommerce-germanized-dhl ); then
+      noroot wp plugin activate woocommerce-germanized-dhl
+    fi
+
+    if $(noroot wp plugin is-installed woocommerce-trusted-shops ); then
+      noroot wp plugin activate woocommerce-trusted-shops
+    fi
+  fi
 }
 
 copy_nginx_configs() {
